requirements.txt
streamlit
pandas
openpyxl
python-dotenv
google-generativeai
langchain
langchain-google-genai